move syntaxHighlighter to REPL package since thats where its used.
