+++ repl/repl.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io"
 	"monkey/evaluator"
-	"monkey/helper"
 	"monkey/lexer"
 	"monkey/object"
 	"monkey/parser"
@@ -36,7 +35,7 @@ func loadBuiltInMethods(env *object.Environment) {
 
 func Start(in io.Reader, out io.Writer) {
 	scanner, env := setup(in, out)
-
+	history := []string{}
 	// Loop forever, until we exit
 	for {
 		fmt.Printf("%s", PROMPT)
@@ -48,6 +47,7 @@ func Start(in io.Reader, out io.Writer) {
 		}
 		// Grab the line we just read
 		line := scanner.Text()
+		history = append(history, line)
 
 		// Exit
 		if line == TERMINATOR {
@@ -70,7 +70,7 @@ func Start(in io.Reader, out io.Writer) {
 		evaluated := evaluator.Eval(program, env)
 		if evaluated != nil {
 			// apply syntax highlighting
-			str := helper.ApplyColorToText(evaluated.Inspect())
+			str := ApplyColorToText(evaluated.Inspect())
 			io.WriteString(out, str)
 			io.WriteString(out, "\n")
 		}
