+++ repl/syntaxHighlighter.go
@@ -1,4 +1,4 @@
-package helper
+package repl
 
 import (
 	"bytes"
